requirements.txt
pymodbus
pyserial
ruamel.yaml
numpy
pandas
loguru
PyQt6